+++ pipelines/drpQA.yaml
@@ -6,3 +6,5 @@ tasks:
     class: pfs.drp.qa.dmCombinedResiduals.DetectorMapCombinedResidualsTask
   extractionQa:
     class: pfs.drp.qa.extractionQa.ExtractionQaTask
+  fluxCalQa:
+    class: pfs.drp.qa.fluxCalQa.FluxCalQaTask
